BackEnd/requirements.txt
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
Deprecated==1.2.13
dnspython==2.2.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Mail==0.9.1
gcloud==0.18.3
googleapis-common-protos==1.56.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.1.0
oauth2client==4.1.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
pymongo==4.0.2
pyparsing==3.0.7
Pyrebase4==4.5.0
python-dotenv==0.19.2
python-jwt==3.3.2
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.14.0
